Split export_table.c into three files
For clarity, this moves the functions in export_table.c into three
separate files.
